+++ src/main/java/com/openclassrooms/climbing/dao/TopoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.openclassrooms.climbing.entities.Topo;
 
-public interface TopoRepository extends JpaRepository <Topo,Integer> {
+public interface TopoRepository extends JpaRepository <Topo,Long> {
 
 }
